requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
distlib==0.3.6
dnspython==2.4.2
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.42.0
frozenlist==1.4.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
langchain==0.0.268
langsmith==0.0.25
loguru==0.7.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.24.4
openai==0.27.8
packaging==23.1
Pillow==10.0.0
pinecone-client==2.2.2
platformdirs==3.8.0
pydantic==2.2.1
pydantic_core==2.6.1
pyparsing==3.0.9
pypdf==3.15.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
tenacity==8.2.3
tiktoken==0.4.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
virtualenv==20.23.1
Werkzeug==2.3.7
win32-setctime==1.1.0
xmltodict==0.13.0
yarl==1.9.2
zipp==3.16.2
